Renaming "app" folder in "rb_backend"
just changed names in imports adn dockerfiles
