+++ backend/dockerfiles/dev
@@ -1,13 +1,13 @@
 FROM radiobretzel/backend:flask
 
-WORKDIR /usr/src/radiobretzel
+WORKDIR /usr/src/rb_backend
 
 COPY requirements-dev.txt requirements-dev.txt
 
 RUN pip install --upgrade pip && \
    pip install -r requirements-dev.txt
 
-VOLUME /usr/src/radiobretzel /var/run/docker.sock
+VOLUME /usr/src/rb_backend /var/run/docker.sock
 EXPOSE 5000
 
 ENV RADIO_BRETZEL_ENV development
