+++ backend/dockerfiles/flask
@@ -5,13 +5,13 @@ LABEL docker.image.name="radiobretzel/backend:flask"
 LABEL docker.image.version="0.1"
 LABEL docker.image.maintainer="Papy Bretzel <papybretzel@ntymail.com>"
 
-WORKDIR /usr/src/radiobretzel
+WORKDIR /usr/src/rb_backend
 COPY hello.py hello.py
 
 RUN pip install --upgrade pip && \
   pip install Flask
 
-VOLUME /usr/src/radiobretzel
+VOLUME /usr/src/rb_backend
 
 #ENTRYPOINT ["/usr/local/bin/python"]
 CMD ["./hello.py"]
