+++ backend/rb_backend/database.py
@@ -1,7 +1,7 @@
 from flask import current_app as app
 from flask_pymongo import PyMongo
 
-from app.errors import DatabaseError
+from rb_backend.errors import DatabaseError
 
 models = [
    'channel',
