+++ backend/rb_backend/docker.py
@@ -1,7 +1,7 @@
 import docker, time
 
 from flask import current_app as app
-from app.errors import DockerError
+from rb_backend.errors import DockerError
 
 def get_docker_client():
    """ Returns an instance of docker client
